feat(laya-metal): split-K for small-m GEMMs, default on; download_into resolves to the most recent write (reflex issue 020 T11)
At the arena's tetris spot sentence (seq 54, 1 noul question) sgemm was
83% of GPU time: the d×d projections ran narrow on a 16×2 = 32
threadgroup grid on a 40-core GPU, each TG walking all of k alone.

- sgemm_splitk + splitk_reduce: batch-1 narrow GEMMs whose grid is
  ≤ 64 narrow TGs are sliced along k in fixed 128-element slices
  (SPLITK_KC) and summed in ascending slice order. Default ON;
  LAYA_METAL_SPLITK=0 turns it off, LAYA_METAL_SPLITK_MAXTGS sets the
  ceiling, Metal::with_splitk is the per-instance seam.
- Packed ≡ loop stays bit-identical: a split row depends only on its own
  data (fixed slice length, fixed reduce order), and the packed encoder
  passes its per-question row segments (Backend::set_row_segments,
  default no-op) so each segment gets the split decision the loop would
  give it. The alternative, making every instance fold at KC, was built
  and measured at +18…+31% on unsplit m ≥ 188 shapes (register
  pressure), and rejected.
- download_into picks the most recently TOUCHED chain entry at a base
  pointer instead of the tightest container. The tightest rule could
  serve a dead key: measured as the packed path's CLS row read from a
  dead 9984-float key (act [0.59, 0.43] vs the loop's [1.0, 0.0]), in
  7/30 processes with split on; 6/6 failing processes had the
  two-candidate download, 0/24 passing ones did. Now 0/30.
- LAYA_METAL_PROFILE=1: opt-in per-dispatch GPU timing (own CB per
  dispatch, GPUStartTime/GPUEndTime), profile_take().

Paired in-process A/B (tests/metal_splitk_ab.rs, 24 position-balanced
rounds, english geometry, M3 Max, background GPU ~22%): on/off median
seq 24 0.615 · 46 0.785 · 54 0.787 · 80 0.878 · 106 0.938 (24/24 wins
each); seq 140–512 1.000 (identical kernels). Cross-binary vs HEAD's
kernel: large-m unchanged within ±1%. End to end, paired per spot
against the torch MPS reference on 372 arena sentences: Rust/Python
1.370 → 1.085.

Gates: G5 (reflex laya_riir_parity) 3/3 checkpoints top-1 1.0, worst
prob drift 7.5e-6; laya_batch_parity green; packed_same_shape_gate
0/30 across processes; metal_ops_smoke 10/10, incl. split vs CPU and vs
unsliced, split row-locality across m (proven to fire: 49056 rows
differ under m-dependent slicing), and deterministic download
resolution (fires under the old rule, diverged 38.7);
packed_forward_equiv; lib 41. clippy -D warnings at laya-riir,
laya-riir-metal, laya-riir-cuda (the last on macOS). Also fixes a
pre-existing unnecessary-cast lint in examples/sgemm_roofline.rs.

Session: katgpt-rs-6c43, 1790340000

+++ crates/riir-infer-laya/examples/sgemm_roofline.rs
@@ -542,7 +542,7 @@ fn main() {
             (rb.len() * 4) as u64,
             RESOURCE_OPTIONS,
         );
-        let rout_buf = device.new_buffer((pm * pn * 4) as u64, RESOURCE_OPTIONS);
+        let rout_buf = device.new_buffer(pm * pn * 4, RESOURCE_OPTIONS);
         let reps = (k / 64) as u32;
         let pn32 = pn as u32;
